kernel: Use uint64 for seconds in datetime conversion routines.

+++ headers/private/kernel/arch/m68k/arch_platform.h
@@ -50,7 +50,7 @@ public:
 	// mimic the PC CMOS
 	virtual uint8 ReadRTCReg(uint8 reg) = 0;
 	virtual void WriteRTCReg(uint8 reg, uint8 val) = 0;
-	virtual	void SetHardwareRTC(uint32 seconds) = 0;
+	virtual	void SetHardwareRTC(uint64 seconds) = 0;
 	virtual	uint32 GetHardwareRTC() = 0;
 
 	virtual void SetHardwareTimer(bigtime_t timeout) = 0;
