+++ src/system/kernel/arch/ppc/arch_platform.cpp
@@ -57,7 +57,7 @@ public:
 	virtual char SerialDebugGetChar();
 	virtual void SerialDebugPutChar(char c);
 
-	virtual	void SetHardwareRTC(uint32 seconds);
+	virtual	void SetHardwareRTC(uint64 seconds);
 	virtual	uint32 GetHardwareRTC();
 
 	virtual	void ShutDown(bool reboot);
@@ -183,7 +183,7 @@ PPCOpenFirmware::SerialDebugPutChar(char c)
 
 
 void
-PPCOpenFirmware::SetHardwareRTC(uint32 seconds)
+PPCOpenFirmware::SetHardwareRTC(uint64 seconds)
 {
 	struct tm t;
 	rtc_secs_to_tm(seconds, &t);
@@ -246,7 +246,7 @@ public:
 	virtual char SerialDebugGetChar();
 	virtual void SerialDebugPutChar(char c);
 
-	virtual	void SetHardwareRTC(uint32 seconds);
+	virtual	void SetHardwareRTC(uint64 seconds);
 	virtual	uint32 GetHardwareRTC();
 
 	virtual	void ShutDown(bool reboot);
@@ -331,7 +331,7 @@ PPCUBoot::SerialDebugPutChar(char c)
 
 
 void
-PPCUBoot::SetHardwareRTC(uint32 seconds)
+PPCUBoot::SetHardwareRTC(uint64 seconds)
 {
 }
